[iOS] Refactor: PostListRequestDTO 수정 및 GetPostsQueryDTO 삭제
