+++ iOS/Village/Village/Domain/Network/PostListRequestDTO.swift
@@ -9,6 +9,23 @@ import Foundation
 
 struct PostListRequestDTO: Encodable {
     
-    let page: Int
+    let searchKeyword: String?
+    let requestFilter: String?
+    let writer: String?
+    let page: String?
+    
+    init(searchKeyword: String? = nil, requestFilter: String? = nil, writer: String? = nil, page: String? = nil) {
+        self.searchKeyword = searchKeyword
+        self.requestFilter = requestFilter
+        self.writer = writer
+        self.page = page
+    }
+    
+    enum CodingKeys: String, CodingKey {
+        case searchKeyword
+        case requestFilter
+        case writer
+        case page
+    }
     
 }
